Move current implementation to cmd/tendermint/commands/debug
